feat: improve reward

+++ python-communication-test/harry_and_thieves_a2c.py
@@ -15,7 +15,7 @@ import collections
 import numpy as np
 import os
 os.environ['TF_XLA_FLAGS'] = '--tf_xla_enable_xla_devices'
-os.environ['CUDA_VISIBLE_DEVICES'] = '-1'
+#os.environ['CUDA_VISIBLE_DEVICES'] = '-1'
 
 """
 def handler(signum, frame):
@@ -52,7 +52,7 @@ class Critic(tf.keras.Model):
 
 
 class ActorCritic():
-    def __init__(self, optimizer, num_actions, num_hidden_units,actions_range):
+    def __init__(self, optimizer, num_actions, num_hidden_units, actions_range):
 
         self.actor_model = Actor(num_actions, num_hidden_units)
         self.critic_model = Critic(num_hidden_units)
@@ -60,12 +60,12 @@ class ActorCritic():
         self.optimizer = optimizer
         self.huber_loss = tf.keras.losses.Huber(
             reduction=tf.keras.losses.Reduction.SUM)
-        
+
         self.actions_range = actions_range
 
         self.clear()
 
-    def predict(self, actor_state, fully_state):
+    def predict(self, actor_state, fully_state, debug=False):
 
         actor_state = tf.expand_dims(actor_state, 0)
         fully_state = tf.expand_dims(fully_state, 0)
@@ -85,13 +85,16 @@ class ActorCritic():
                     action_probility_logits = tf.nn.softmax(action_probility)
                     start_index += action
 
-                    actions.append(self.get_action(tf.expand_dims(action_probility,0))[0,0])
+                    actions.append(self.get_action(
+                        tf.expand_dims(action_probility, 0))[0, 0])
                     prod = prod * action_probility_logits[actions[-1]]
 
                 critic_predict = self.critic_model(fully_state)
 
                 self.actor_predicts.append(prod)
                 self.critic_predicts.append(critic_predict[0, 0])
+                if debug:
+                    print(prod, critic_predict[0, 0])
 
                 return actions, critic_predict[0, 0]
 
@@ -124,14 +127,15 @@ class ActorCritic():
             with self.critic_tape:
 
                 eps_len = len(rewards)
-                expected_rewards = self.get_expected_rewards(rewards)
+                expected_rewards = self.get_expected_return(rewards)
                 advantage = self.advantage_compute(
-                    self.critic_predicts, rewards)
-                
+                    self.critic_predicts, expected_rewards)
+
                 actions_log = self.actions_log_compute(self.actor_predicts)
                 actor_loss = -tf.reduce_sum(actions_log * advantage)
 
-                critic_loss = self.huber_loss(self.critic_predicts, rewards)
+                critic_loss = self.huber_loss(
+                    self.critic_predicts, expected_rewards)
                 return actor_loss, critic_loss
 
     def get_expected_rewards(self, reward):
@@ -143,6 +147,31 @@ class ActorCritic():
             reward_sum -= reward[i]
 
         return exp_reward
+
+    def get_expected_return(self, rewards: tf.Tensor, gamma: float = .99, standardize: bool = True) -> tf.Tensor:
+        """Compute expected returns per timestep."""
+        eps = np.finfo(np.float32).eps.item()
+
+        n = tf.shape(rewards)[0]
+        returns = tf.TensorArray(dtype=tf.float32, size=n)
+
+        # Start from the end of `rewards` and accumulate reward sums
+        # into the `returns` array
+        rewards = tf.cast(rewards[::-1], dtype=tf.float32)
+        discounted_sum = tf.constant(0.0)
+        discounted_sum_shape = discounted_sum.shape
+        for i in tf.range(n):
+            reward = rewards[i]
+            discounted_sum = reward + gamma * discounted_sum
+            discounted_sum.set_shape(discounted_sum_shape)
+            returns = returns.write(i, discounted_sum)
+        returns = returns.stack()[::-1]
+
+        if standardize:
+            returns = ((returns - tf.math.reduce_mean(returns)) /
+                       (tf.math.reduce_std(returns) + eps))
+
+        return returns
 
     def advantage_compute(self, values, rewards):
         advantage = []
@@ -167,38 +196,41 @@ class ActorCritic():
                 self.critic_predicts = []
 
 
-num_hidden_units = 512
+num_hidden_units = 2048
 
 
 class Worker():
-    def __init__(self, port):
+    def __init__(self, port, subprocess=True):
 
-        args = ['unity_game/HarryAndTheStone.exe', '--p', str(port)]
-        self.subprocess = subprocess.Popen(args)
+        if subprocess:
+            args = ['unity_game/HarryAndTheStone.exe', '--p', str(port)]
+            self.subprocess = subprocess.Popen(args)
+
         self.env = Unity(port)
 
-        optimizer = tf.keras.optimizers.Adam(learning_rate=0.005)
+        optimizer = tf.keras.optimizers.Adam(learning_rate=5e-4)
 
-        self.harry_a2c = ActorCritic(optimizer, 4, num_hidden_units, [2, 2])
+        self.harry_a2c = ActorCritic(optimizer, 6, num_hidden_units, [3, 3])
 
         self.first_thieve_a2c = ActorCritic(
-            optimizer, 6, num_hidden_units, [2, 2, 2])
+            optimizer, 8, num_hidden_units, [3, 3, 2])
         self.second_thieve_a2c = ActorCritic(
-            optimizer, 6, num_hidden_units, [2, 2, 2])
+            optimizer, 8, num_hidden_units, [3, 3, 2])
         """
         self.first_thieve_a2c_grab = ActorCritic(
             optimizer, 2, num_hidden_units)
         self.second_thieve_a2c_grab = ActorCritic(
             optimizer, 2, num_hidden_units)
         """
+
     def apply_harry(self, frame_action, harry_state, fully_state):
 
         actions, _ = self.harry_a2c.predict(harry_state, fully_state)
-        
-        x = 2 * int(actions[0]) - 1
-        y = 2 * int(actions[1]) - 1
 
-        self.env.apply_harry_action(frame_action, x,y)
+        x = int(actions[0]) - 1
+        y = int(actions[1]) - 1
+
+        self.env.apply_harry_action(frame_action, x, y)
 
     def grab_converter(self, grab):
         if grab == 0:
@@ -210,15 +242,14 @@ class Worker():
 
         if thieve_id == 0:
             actions, _ = self.first_thieve_a2c.predict(
-                thieve_state, fully_state)
+                thieve_state, fully_state, False)
         else:
             actions, _ = self.second_thieve_a2c.predict(
                 thieve_state, fully_state)
 
-        x = 2 * int(actions[0]) - 1
-        y = 2 * int(actions[1]) - 1
+        x = int(actions[0]) - 1
+        y = int(actions[1]) - 1
         grab = int(actions[2])
-
         if thieve_id == 0:
             self.env.apply_first_thieve_action(frame_action, x, y, grab)
         else:
@@ -227,9 +258,10 @@ class Worker():
     def run_episode(self):
         state = self.env.reset()
         harry_rewards = []
-        thieve_rewards = []
+        first_thieve_rewards = []
+        second_thieve_rewards = []
 
-        first_thieve_a2c_done = False
+        first_thieve_done = False
         second_thieve_done = False
 
         while(True):
@@ -240,36 +272,40 @@ class Worker():
 
             self.apply_harry(frame_action, harry_state, full_state)
 
-            if not first_thieve_a2c_done:
-                self.apply_thieve(frame_action, 0, fir_thieve_state, full_state)
+            if not first_thieve_done:
+                self.apply_thieve(
+                    frame_action, 0, fir_thieve_state, full_state)
 
             if not second_thieve_done:
-                self.apply_thieve(frame_action, 1, sec_thieve_state, full_state)
+                self.apply_thieve(
+                    frame_action, 1, sec_thieve_state, full_state)
 
             state, rewards, done = self.env.action(frame_action)
 
+            if not first_thieve_done:
+                first_thieve_rewards.append(rewards[1])
+
+            if not second_thieve_done:
+                second_thieve_rewards.append(rewards[2])
+
             harry_rewards.append(rewards[0])
-            thieve_rewards.append(rewards[1])
 
             if done[0]:
-                return harry_rewards, thieve_rewards
+                return harry_rewards, first_thieve_rewards, second_thieve_rewards
 
             if done[1]:
-                first_thieve_a2c_done = True
+                first_thieve_done = True
 
             if done[2]:
                 second_thieve_done = True
 
     def train(self, max_ep):
         for i in range(max_ep):
-            harry_reward, thieve_reward = self.run_episode()
+            harry_reward, first_thieve_rewards, second_thieve_rewards = self.run_episode()
 
             self.harry_a2c.apply_loss(harry_reward)
-            self.first_thieve_a2c.apply_loss(thieve_reward)
-            self.second_thieve_a2c.apply_loss(thieve_reward)
-            
-            #self.first_thieve_a2c_grab.apply_loss(thieve_reward)
-            #self.second_thieve_a2c_grab.apply_loss(thieve_reward)
+            self.first_thieve_a2c.apply_loss(first_thieve_rewards)
+            self.second_thieve_a2c.apply_loss(second_thieve_rewards)
 
             global eps_count
             eps_count += 1
@@ -278,12 +314,12 @@ class Worker():
         self.close()
 
     def close(self):
-        self.subprocess.terminate()
+        if subprocess:
+            self.subprocess.terminate()
 
 
-seed = 156
-tf.random.set_seed(seed)
-np.random.seed(seed)
+#tf.random.set_seed(seed)
+#np.random.seed(seed)
 
 print("Start learning")
 
@@ -293,6 +329,10 @@ running_reward = 0
 
 
 def main():
+
+    worker = Worker(7979, subprocess=False)
+    worker.train(100000)
+
     import time
     t = time.time()
